Move the core hacks to their own file, re-enable the notifications tests, and shift away another helper method that's shared
