+++ modules/shared/php/helper-utilities.php
@@ -29,4 +29,21 @@ class HelperUtilities {
 
 		return $post_types;
 	}
+
+	/**
+	 * Whether or not the current page is our settings view. Determination is based on $pagenow, $_GET['page'], and if it's settings module or not.
+	 *
+	 * @return bool $is_settings_view Return true if it is
+	 */
+	public static function is_settings_view_loaded( string $slug ): bool {
+		global $pagenow;
+
+		// All of the settings views are based on admin.php and a $_GET['page'] parameter
+		if ( 'admin.php' != $pagenow || ! isset( $_GET['page'] ) ) {
+			return false;
+		}
+
+		// The current page better be in the array of registered settings view slugs
+		return $_GET['page'] === $slug;
+	}
 }
