+++ tests/modules/custom-status/meta/test-required-metadata-id-handler.php
@@ -26,7 +26,6 @@ class RequiredMetadataIdHandlerTest extends WP_UnitTestCase {
 		CustomStatus::setup_install();
 	}
 
-	// tear down after class
 	protected function tearDown(): void {
 		parent::tearDown();
