+++ tests/modules/custom-status/meta/test-required-user-id-handler.php
@@ -26,7 +26,6 @@ class RequiredUserIdHandlerTest extends WP_UnitTestCase {
 		CustomStatus::setup_install();
 	}
 
-	// tear down after class
 	protected function tearDown(): void {
 		parent::tearDown();
