+++ tests/modules/notifications/test-notifications.php
@@ -9,6 +9,8 @@
 namespace VIPWorkflow\Tests;
 
 use VIPWorkflow\Modules\Notifications;
+use VIPWorkflow\Modules\Settings;
+use VIPWorkflow\Modules\Shared\PHP\OptionsUtilities;
 use WP_Error;
 use WP_UnitTestCase;
 
@@ -18,6 +20,9 @@ class NotificationsTest extends WP_UnitTestCase {
 		parent::tearDown();
 
 		reset_phpmailer_instance();
+
+		// Reset all module options
+		OptionsUtilities::reset_all_module_options();
 	}
 
 	public function test_validate_get_notification_footer() {
@@ -43,41 +48,41 @@ class NotificationsTest extends WP_UnitTestCase {
 		$this->assertDiscardWhitespace( $body, $email->body );
 	}
 
-	// public function test_send_to_webhook_happy_path() {
-	//  // Hook in and return a known response
-	//  add_filter( 'pre_http_request', function () {
-	//      return array(
-	//          'headers'     => array(),
-	//          'cookies'     => array(),
-	//          'filename'    => null,
-	//          'response'    => 200,
-	//          'status_code' => 200,
-	//          'success'     => 1,
-	//          'body'        => 'All Done',
-	//      );
-	//  }, 10, 3 );
+	public function test_send_to_webhook_happy_path() {
+		// Hook in and return a known response
+		add_filter( 'pre_http_request', function () {
+			return array(
+				'headers'     => array(),
+				'cookies'     => array(),
+				'filename'    => null,
+				'response'    => 200,
+				'status_code' => 200,
+				'success'     => 1,
+				'body'        => 'All Done',
+			);
+		}, 10, 3 );
 
-	//  VIP_Workflow::instance()->settings->module->options->webhook_url = 'https://webhook.site/this-url-doesnt-exist';
+		OptionsUtilities::update_module_option_key( Settings::SETTINGS_SLUG, 'webhook_url', 'https://webhook.site/this-url-doesnt-exist' );
 
-	//  $response = Notifications::send_to_webhook( 'Test Message', 'status-change', '2024-09-19 00:26:50' );
+		$response = Notifications::send_to_webhook( 'Test Message', 'status-change', '2024-09-19 00:26:50' );
 
-	//  $this->assertTrue( $response );
+		$this->assertTrue( $response );
 
-	//  VIP_Workflow::instance()->settings->module->options->webhook_url = '';
-	// }
+		OptionsUtilities::update_module_option_key( Settings::SETTINGS_SLUG, 'webhook_url', '' );
+	}
 
-	// public function test_send_to_webhook_error_path() {
-	//  // Hook in and return a known response
-	//  add_filter( 'pre_http_request', function () {
-	//      return new WP_Error( 'http_request_failed', 'Error Message' );
-	//  }, 10, 3 );
+	public function test_send_to_webhook_error_path() {
+		// Hook in and return a known response
+		add_filter( 'pre_http_request', function () {
+			return new WP_Error( 'http_request_failed', 'Error Message' );
+		}, 10, 3 );
 
-	//  VIP_Workflow::instance()->settings->module->options->webhook_url = 'https://webhook.site/this-url-doesnt-exist';
+		OptionsUtilities::update_module_option_key( Settings::SETTINGS_SLUG, 'webhook_url', 'https://webhook.site/this-url-doesnt-exist' );
 
-	//  $response = Notifications::send_to_webhook( 'Test Message', 'status-change', '2024-09-19 00:26:50' );
+		$response = Notifications::send_to_webhook( 'Test Message', 'status-change', '2024-09-19 00:26:50' );
 
-	//  $this->assertFalse( $response );
+		$this->assertFalse( $response );
 
-	//  VIP_Workflow::instance()->settings->module->options->webhook_url = '';
-	// }
+		OptionsUtilities::update_module_option_key( Settings::SETTINGS_SLUG, 'webhook_url', '' );
+	}
 }
